+++ external-table/sql/pxf.sql
@@ -8,9 +8,9 @@ SELECT * from pxf_readcustom_test order by a;
 ------------------------------------------------------------------
 -- PXF write test
 ------------------------------------------------------------------
-\!rm -rf /tmp/pxf
+\! rm -rf /tmp/pxf
 
 INSERT INTO pxf_write_test SELECT * from origin;
 
-\!ls -1 /tmp/pxf/ | wc -l | sed -e 's/^[[:space:]]*//'
-\!cat /tmp/pxf/* | sort
+\! ls -1 /tmp/pxf/ | wc -l | sed -e 's/^[[:space:]]*//'
+\! cat /tmp/pxf/* | sort