+++ external-table/src/pxffilters.c
@@ -24,6 +24,7 @@
  */
 #include "pxffilters.h"
 
+#include "catalog/pg_type.h"
 #include "catalog/pg_operator.h"
 #include "optimizer/clauses.h"
 #include "parser/parse_expr.h"
@@ -176,7 +177,7 @@ dbop_pxfop_map pxf_supported_opr_op_expr[] =
 	{85 /* boolne */ , PXFOP_NE},
 
 	/* bpchar */
-	{BPCharEqualOperator /* bpchareq */ , PXFOP_EQ},
+	{1054 /* bpchareq */ , PXFOP_EQ},
 	{1058 /* bpcharlt */ , PXFOP_LT},
 	{1060 /* bpchargt */ , PXFOP_GT},
 	{1059 /* bpcharle */ , PXFOP_LE},
@@ -238,7 +239,7 @@ dbop_pxfop_array_map pxf_supported_opr_scalar_array_op_expr[] =
 	{1120 /* float48eq */ , PXFOP_IN, true},
 
 	/* bpchar */
-	{BPCharEqualOperator /* bpchareq */ , PXFOP_IN, true},
+	{1054 /* bpchareq */ , PXFOP_IN, true},
 };